requirements.txt
jpype1
mpxj
django
djangorestframework